@@ -1,66 +1,54 @@
 {
-  "name": "frontend",
-  "version": "2.0.0",
+  "name": "contacts",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --proxy-config proxy.conf.json --host 0.0.0.0 --watch --poll 2000",
+    "start": "ng serve",
     "serve": "ng serve --proxy-config proxy.conf.json",
-    "build": "ng build --configuration production",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "dev:ssr": "ng run frontend:serve-ssr",
-    "serve:ssr": "node dist/frontend/server/main.js",
-    "build:prod": "ng build --configuration production --output-path=dist",
-    "build:ssr": "ng build --configuration production && ng run frontend:server:production",
-    "build:output": "npm run build:ssr --output-path=dist",
-    "prerender": "ng run frontend:prerender",
-    "dev-server": "concurrently \"nodemon ../api/server.js --watch api\" \"npm run serve\""
+    "serve:ssr:contacts": "node dist/contacts/server/server.mjs"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.3",
-    "@angular/common": "^17.0.3",
-    "@angular/compiler": "^17.0.3",
-    "@angular/core": "^17.0.3",
-    "@angular/forms": "^17.0.3",
-    "@angular/platform-browser": "^17.0.3",
-    "@angular/platform-browser-dynamic": "^17.0.3",
-    "@angular/platform-server": "^17.0.3",
-    "@angular/router": "^17.0.3",
-    "@angular/ssr": "^17.0.1",
-    "@nguniversal/express-engine": "^16.2.0",
-    "@swimlane/ngx-datatable": "^20.0.0",
-    "bootstrap": "^4.5.3",
-    "express": "^4.15.2",
-    "font-awesome": "^4.7.0",
-    "ngx-bootstrap": "^11.0.2",
-    "ngx-toastr": "^17.0.2",
-    "rxjs": "^6.6.7",
-    "tslib": "^2.3.1"
+    "@angular/animations": "^18.1.0",
+    "@angular/common": "^18.1.0",
+    "@angular/compiler": "^18.1.0",
+    "@angular/core": "^18.1.0",
+    "@angular/forms": "^18.1.0",
+    "@angular/platform-browser": "^18.1.0",
+    "@angular/platform-browser-dynamic": "^18.1.0",
+    "@angular/platform-server": "^18.1.0",
+    "@angular/router": "^18.1.0",
+    "@angular/ssr": "^18.1.0",
+    "@fortawesome/fontawesome-free": "^6.5.2",
+    "@ng-bootstrap/ng-bootstrap": "^17.0.0",
+    "@popperjs/core": "^2.11.8",
+    "bootstrap": "^5.3.2",
+    "express": "^4.18.2",
+    "ngx-toastr": "^19.0.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.1",
-    "@angular/cli": "^17.0.1",
-    "@angular/compiler-cli": "^17.0.3",
-    "@angular/language-service": "^17.0.3",
-    "@types/express": "^4.17.20",
-    "@types/jasmine": "^5.1.1",
-    "@types/jasminewd2": "^2.0.12",
-    "@types/node": "^20.8.9",
-    "codelyzer": "^6.0.2",
-    "concurrently": "^8.2.2",
-    "jasmine-core": "^5.1.1",
-    "jasmine-spec-reporter": "^7.0.0",
-    "karma": "^6.4.1",
-    "karma-chrome-launcher": "^3.1.1",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "^5.1.0",
-    "karma-jasmine-html-reporter": "^2.0.0",
-    "protractor": "^7.0.0",
-    "ts-node": "^10.9.1",
-    "tslint": "^6.1.3",
-    "typescript": "^5.2.2",
-    "zone.js": "^0.14.2"
+    "@angular-devkit/build-angular": "^18.1.0",
+    "@angular/cli": "^18.1.0",
+    "@angular/compiler-cli": "^18.1.0",
+    "@angular/localize": "^18.1.0",
+    "@types/express": "^4.17.17",
+    "@types/jasmine": "~5.1.0",
+    "@types/node": "^18.18.0",
+    "jasmine-core": "~5.1.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "prettier": "3.3.0",
+    "prettier-plugin-organize-imports": "3.2.4",
+    "prettier-plugin-tailwindcss": "0.6.1",
+    "typescript": "~5.4.2"
   }
 }
