@@ -1,50 +1,58 @@
-import { Injectable } from "@angular/core";
-import { HttpClient, HttpErrorResponse } from "@angular/common/http";
+import { HttpClient, HttpErrorResponse } from '@angular/common/http';
+import { Injectable } from '@angular/core';
 
-import { map, catchError } from "rxjs/operators";
+import { catchError, map } from 'rxjs/operators';
 
-import { environment } from "../../../environments/environment";
-import { throwError } from "rxjs";
+import { throwError } from 'rxjs';
+import { environment } from '../../../environments/environment';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class ContactService {
-  constructor(private http: HttpClient) {}
+    constructor(private http: HttpClient) {}
 
-  getAll() {
-    return this.http.get<any[]>(environment.apiEndpoint + "/v1/contact").pipe(
-      map((res: any) => res.data),
-      catchError(this.handleErrorObservable)
-    );
-  }
+    getAll() {
+        return this.http.get<any[]>(environment.apiEndpoint + '/v1/contact').pipe(
+            map((res: any) => res.data),
+            catchError(this.handleErrorObservable)
+        );
+    }
 
-  getById(id: string) {
-    return this.http.get(environment.apiEndpoint + "/v1/contact/" + id).pipe(
-      map((res: any) => res.data),
-      catchError(this.handleErrorObservable)
-    );
-  }
+    getById(_id: string) {
+        return this.http.get(environment.apiEndpoint + '/v1/contact/' + _id).pipe(
+            map((res: any) => res.data),
+            catchError(this.handleErrorObservable)
+        );
+    }
 
-  create(contact: any) {
-    return this.http.post(environment.apiEndpoint + "/v1/contact", contact).pipe(
-      map((res: any) => res.data),
-      catchError(this.handleErrorObservable)
-    );
-  }
+    create(contact: any) {
+        return this.http
+            .post(environment.apiEndpoint + '/v1/contact', contact)
+            .pipe(
+                map((res: any) => res.data),
+                catchError(this.handleErrorObservable)
+            );
+    }
 
-  update(contact: any) {
-    return this.http.put(environment.apiEndpoint + "/v1/contact/" + contact.id, contact).pipe(
-      map((res: any) => res.data),
-      catchError(this.handleErrorObservable)
-    );
-  }
+    update(contact: any) {
+        return this.http
+            .put(environment.apiEndpoint + '/v1/contact/' + contact.id, contact)
+            .pipe(
+                map((res: any) => res.data),
+                catchError(this.handleErrorObservable)
+            );
+    }
 
-  delete(id: string) {
-    return this.http.delete(environment.apiEndpoint + "/v1/contact/" + id).pipe(
-      map((res: any) => res.data),
-      catchError(this.handleErrorObservable)
-    );
-  }
-  private handleErrorObservable(error: HttpErrorResponse) {
-    return throwError(error);
-  }
+    delete(_id: string) {
+        return this.http
+            .delete(environment.apiEndpoint + '/v1/contact/' + _id)
+            .pipe(
+                map((res: any) => res.data),
+                catchError(this.handleErrorObservable)
+            );
+    }
+    private handleErrorObservable(error: HttpErrorResponse) {
+        return throwError(error);
+    }
 }
